requirements.txt
django
Pillow
django-rest-auth